Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
python-dotenv
passlib
python-jose
python-multipart